requirements.txt
python-dotenv
azure-identity